純粋関数を別ファイルに切り出しvi.mockを全廃
groupTagsByBillIdをshared/utils/group-tags.tsに、
collectAskedQuestionIdsをshared/collect-asked-question-ids.tsに
それぞれ切り出し、テストからserver-onlyのmockを不要にした。
