+++ web/src/features/bills/server/repositories/bill-repository.ts
@@ -144,22 +144,7 @@ export async function findBillContentByDifficulty(
 // Tags (bulk)
 // ============================================================
 
-type BillTag = {
-  bill_id: string;
-  tags: { id: string; label: string } | null;
-};
-
-export function groupTagsByBillId(
-  billTags: BillTag[]
-): Map<string, Array<{ id: string; label: string }>> {
-  return billTags.reduce((acc, bt) => {
-    if (bt.tags) {
-      const existing = acc.get(bt.bill_id) ?? [];
-      acc.set(bt.bill_id, [...existing, bt.tags]);
-    }
-    return acc;
-  }, new Map<string, Array<{ id: string; label: string }>>());
-}
+import { groupTagsByBillId } from "../../shared/utils/group-tags";
 
 /**
  * 複数のbill_idに紐づくタグを一括取得し、bill_idごとにグループ化して返す
